# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend-API fuer das Schulbibliothek-Verwaltungssystem.",
        "title": "Schulbibliothek API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/klassen": {
            "get": {
                "description": "Retrieves all unique school class names currently assigned to students.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get list of classes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schueler": {
            "get": {
                "description": "Retrieves students, optionally filtered by a specific school class, along with loan counts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "List students",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School class to filter by",
                        "name": "klasse",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new student profile in the library database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Create student",
                "parameters": [
                    {
                        "description": "Student creation payload",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schueler/{id}": {
            "get": {
                "description": "Retrieves the complete profile for a student by their ID, including active loans and avatar photo URL if present.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get student profile details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StudentProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Transactionally deletes a student from the system, checks for active loans or unpaid damage fees, anonymizes historical loans, and writes to audit_log.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Delete student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.BorrowedBook": {
            "type": "object",
            "properties": {
                "ausgeliehen_am": {
                    "type": "string"
                },
                "autor": {
                    "type": "string"
                },
                "barcode_id": {
                    "type": "string"
                },
                "cover_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rueckgabe_frist": {
                    "type": "string"
                },
                "titel": {
                    "type": "string"
                }
            }
        },
        "api.CreateStudentRequest": {
            "type": "object",
            "properties": {
                "barcode_id": {
                    "type": "string"
                },
                "klasse": {
                    "type": "string"
                },
                "nachname": {
                    "type": "string"
                },
                "vorname": {
                    "type": "string"
                }
            }
        },
...
        "api.StudentProfileResponse": {
            "type": "object",
            "properties": {
                "abgaenger_jahr": {
                    "type": "integer"
                },
                "barcode_id": {
                    "type": "string"
                },
                "entliehene_buecher": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BorrowedBook"
                    }
                },
                "foto_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ist_gesperrt": {
                    "type": "boolean"
                },
                "klasse": {
                    "type": "string"
                },
                "nachname": {
                    "type": "string"
                },
                "vorname": {
                    "type": "string"
                }
            }
        },
...
    }
}